+++ psipy/tracing/flines.py
@@ -12,33 +12,38 @@ __all__ = ["FieldLines", "FieldLine"]
 class FieldLine:
     """
     A single field line.
+
+    Parameters
+    ----------
+    r : numpy.ndarray
+        Radial coordinates.
+    lat : numpy.ndarray
+        Latitude coordinates **in radians**.
+    lon : numpy.ndarray
+        Longitude coordinates **in radians**.
+    runit : astropy.units.Unit
+        Radial coordinate unit.
     """
 
-    r: np.ndarray
-    lat: np.ndarray
-    lon: np.ndarray
+    r: u.Quantity
+    lat: u.Quantity
+    lon: u.Quantity
 
-    def __init__(self, *, r, lat, lon):
-        """
-        Parameters
-        ----------
-        r : numpy.ndarray
-            Radial coordinates.
-        lat : numpy.ndarray
-            Latitude coordinates **in radians**.
-        lon : numpy.ndarray
-            Longitude coordinates **in radians**.
-        """
-        self.r = r
+    def __init__(
+        self, *, r: np.ndarray, lat: np.ndarray, lon: np.ndarray, runit: u.Unit
+    ):
+        self.r = r * runit
         self.lat = lat * u.rad
         self.lon = lon * u.rad
+        self.runit = runit
 
     @property
-    def xyz(self):
+    def xyz(self) -> u.Quantity:
         """
         Cartesian coordinates as a (n, 3) shaped array.
         """
-        return np.array(spherical_to_cartesian(self.r, self.lat, self.lon)).T
+        x, y, z = spherical_to_cartesian(self.r, self.lat, self.lon)
+        return np.array([x, y, z]).T * self.r.unit
 
     @property
     def _rlatlon(self):
@@ -46,7 +51,11 @@ class FieldLine:
         Spherical coordinates as a (n, 3) shaped array.
         """
         return np.column_stack(
-            [self.r, self.lat.to_value(u.rad), self.lon.to_value(u.rad)]
+            [
+                self.r.to_value(self.runit),
+                self.lat.to_value(u.rad),
+                self.lon.to_value(u.rad),
+            ]
         )
 
 
@@ -58,15 +67,20 @@ class FieldLines:
 
     flines: List[FieldLine]
 
-    def __init__(self, xs):
+    def __init__(self, xs: np.ndarray, runit: u.Unit):
         """
         Parameters
         ----------
         xs : list[numpy.ndarray]
             Field lines. Each array must have lon, lat, r columns in that
             order.
+        runit : astropy.units.Unit
+            Unit for radial coordinate.
         """
-        self.flines = [FieldLine(r=x[:, 2], lat=x[:, 1], lon=x[:, 0]) for x in xs]
+        self.flines = [
+            FieldLine(r=x[:, 2], lat=x[:, 1], lon=x[:, 0], runit=runit) for x in xs
+        ]
+        self.runit = runit
 
     def __getitem__(self, i):
         return self.flines[i]
@@ -91,7 +105,9 @@ class FieldLines:
         -----
         Arrays are saved using `numpy.savez_compressed`.
         """
-        np.savez_compressed(filename, *[fline._rlatlon for fline in self.flines])
+        flines = {f"fline_{i}": fline._rlatlon for i, fline in enumerate(self.flines)}
+        flines["runit"] = np.array(self.runit.to_string())
+        np.savez_compressed(filename, **flines)
 
     @classmethod
     def load(cls, filename):
@@ -110,4 +126,12 @@ class FieldLines:
         flines : FieldLines
         """
         arrs = np.load(str(filename))
-        return cls([arrs[k][:, ::-1] for k in arrs])
+        if "runit" in arrs:
+            runit = u.Unit(str(arrs["runit"]))
+        else:
+            # For backwards compatability with versions < 0.4, assume
+            # solar radii
+            runit = u.R_sun
+
+        fline_data = np.array([arrs[k][:, ::-1] for k in arrs if k != "runit"])
+        return cls(fline_data, runit=runit)