+++ psipy/tracing/tests/test_tracing.py
@@ -6,11 +6,9 @@ from psipy.model import MASOutput
 from psipy.tracing import FieldLines, FortranTracer
 
 
-def test_tracer():
+def test_tracer(mas_model):
     # Simple smoke test of field line tracing
-    mas_path = sample_data.mas_sample_data()
-    model = MASOutput(mas_path)
-    bs = model.cell_corner_b()
+    bs = mas_model.cell_corner_b()
     # Fake data to be unit vectors pointing in radial direction
     bs.loc[..., "bp"] = 0
     bs.loc[..., "bt"] = 0
@@ -19,7 +17,7 @@ def test_tracer():
     def cell_corner_b(self):
         return bs
 
-    model.cell_corner_b = cell_corner_b
+    mas_model.cell_corner_b = cell_corner_b
 
     tracer = FortranTracer()
 
@@ -27,7 +25,7 @@ def test_tracer():
     lat = 0 * u.deg
     lon = 0 * u.deg
 
-    flines = tracer.trace(model, lon=lon, lat=lat, r=r)
+    flines = tracer.trace(mas_model, lon=lon, lat=lat, r=r)
     assert len(flines) == 1
 
     # Check that with auto step size, number of steps is close to number of
@@ -36,13 +34,13 @@ def test_tracer():
     assert flines[0].xyz.shape == (139, 3)
 
     tracer = FortranTracer(step_size=0.5)
-    flines = tracer.trace(model, lon=lon, lat=lat, r=r)
+    flines = tracer.trace(mas_model, lon=lon, lat=lat, r=r)
     # Check that changing step size has an effect
     assert flines[0].xyz.shape == (278, 3)
 
 
 def test_fline_io(mas_model, tmpdir):
-    # Simple smoke test of field line tracing
+    # Test saving and loading field lines
     mas_path = sample_data.mas_sample_data()
     model = MASOutput(mas_path)
     tracer = FortranTracer()